ICalRecurrenceRule now supports WEEKLY recurrence with the BYDAY attribute.
A common (and relatively simple) recurrence rule is to repeat several days
per week, every week as seen in the example below.

Example:

	BEGIN:VEVENT
	DTSTART;TZID=America/New_York:20110620T073000
	DTEND;TZID=America/New_York:20110621T000000
	RRULE:FREQ=WEEKLY;UNTIL=20110811T113000Z;BYDAY=MO,TU,WE,TH
	DTSTAMP:20110824T134544Z
	UID:o0742l784numhoobgl50sh58f4@google.com
	CREATED:20110504T185446Z
	DESCRIPTION:
	LAST-MODIFIED:20110603T034935Z
	LOCATION:
	SEQUENCE:0
	STATUS:CONFIRMED
	SUMMARY:Regular hours are 7:30 am - midnight
	TRANSP:OPAQUE
	END:VEVENT

Additional recurrence cycles such as BYMONTH, BYWEEKNO, BYMONTHDAY, and
INTERVALs are not supported by this change, but could be implemented
in a similar fashion. For more examples of recurrence cycles see:
    http://www.kanzaki.com/docs/ical/rrule.html