Fixed Multi-site/Watchdog conflict
Multi-site was attempting to use realpath_exists to check the location of
the $siteDir, but Watchdog changed realpath_exists to have extra
semantics.  Fixed Multi-site code to use a combination of realpath and
file_exists to verify location, in order to get around Watchdog's
restrictions.  This code is setting the location of SITE_DIR, which cannot
be used to detect safe locations until it is set.