Multiple Memcache servers supported. Use config arrays.
If an array of hosts are used, but a port string, that port will
be used for all hosts. If an array of ports is given with an
array of hosts, they will be matched by numerial index. If the
hosts array is larger than the ports array, the last element in
the ports array will be used to pad the ports array to the size of
the hosts array. If the ports array is larger than the hosts array
the extra port elements are discarded.