CASAuthentication - Add support for proxy-authenticated web-service requests.
The CAS 2.0 protocol supports proxy-authentication of web-service requests.
In a proxy-authentication configuration, phpCAS with Kurogo is given a
proxy-granting-ticket (PGT) which it can use to obtain a proxy-ticket (PT)
that it passes off to underlying web-services secured by CAS. The underlying
web-service will validate the PT with the CAS server and obtain information
about the user who logged in to Kurogo without ever seeing the user's password.

This patch adds configuration and documentation for the CASAuthentication
authority.