Changed name of Athletics Schedule Parser class in admin-module.json from ICSDataParser to the correct name of ICSAthleticParser.
Added check to see if a Photo publish date had been set and not try to format an empty date. Also, added photo description to the list of properties set for an image from RSS. Modified the photo show template to only attempt to show description if it exists and only show photoDate if it exists.

Added check to HTTPDataResponse.php so that when we determine whether or not there has been an HTTP error, we also make sure that out url started with "http". This was causing errors to be set on the object when some other class (which I can't recall now) was using the HTTPData classes to obtain data from files on the system, rather than files via URL.