Removed break that was causing Maps to display an infowindow by default no matter how many placemarks were added.
Modified Admin module css and js to allow for dynamic groups that extend to more than one line to fit without overlapping content. Add several groups to the Maps app to see this.

Modified the DateFormatter to allow for defining and passing in your own patterns. If the parameter you pass in is non-numeric, then we don't bother looking at the preset constants and instead use whatever pattern was passed in.

Changed GoogleJSMap to get the Description from the KML file since there is no subtitle entry. Also changed several occurrences of misnamed variables.