KGO-1372 Added parameter to disallow map viewport adjusting when adding placemarks in maps
The parameter is called 'resizeOnAddPlacemark' in the MapImageController. The default parameter setting is to allow map viewport to automatically recalculate the map viewport when a placemark is added (the original behavior). I needed the option to turn this off as it can cause the map initialization to ignore zoom and center and is particularly ugly when the map has world-spanning placemarkers.