Allow setting the ACTIVE_SITE via the environment.
This patch allows Kurogo to serve multiple sites from the same code directory
where the active-site selection is configured by something beyond the limited subpath
structure of MultiSite mode.

One example enabled by this patch is to serve different Kurogo sites for each
VirtualHost on the machine:

	<VirtualHost *:80>
		ServerName students.example.edu
		DocumentRoot /var/www/kurogo/www/

		SetEnv ACTIVE_SITE Students
	</VirtualHost>

	<VirtualHost *:80>
		ServerName faculty.example.edu
		DocumentRoot /var/www/kurogo/www/

		SetEnv ACTIVE_SITE Faculty
	</VirtualHost>

This patch does not interfere with normal single-site operation or with MultiSite
mode. See the additions to doc/mw/multisite.rst in this patch for more detailed
configuration instructions.